(PC-32696)[PRO] refactor: mv IndividualOfferBookings to new arch folder
