+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/IndividualOfferBookings/IndividualOfferBookings.tsx
@@ -4,12 +4,13 @@ import React from 'react'
 import { useIndividualOfferContext } from 'commons/context/IndividualOfferContext/IndividualOfferContext'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { ActionBar } from 'components/IndividualOffer/ActionBar/ActionBar'
-import { BookingsSummaryScreen } from 'components/IndividualOffer/BookingsSummary/BookingsSummary'
 import { IndivualOfferLayout } from 'components/IndividualOffer/IndivualOfferLayout/IndivualOfferLayout'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
 import { Spinner } from 'ui-kit/Spinner/Spinner'
 
-const BookingsSummary = (): JSX.Element | null => {
+import { IndividualOfferBooklngsSummaryScreen } from './components/IndividualOfferBooklngsSummaryScreen'
+
+const IndividualOfferBookings = (): JSX.Element | null => {
   const mode = useOfferWizardMode()
   const { offer } = useIndividualOfferContext()
 
@@ -19,7 +20,7 @@ const BookingsSummary = (): JSX.Element | null => {
 
   return (
     <IndivualOfferLayout title="Récapitulatif" offer={offer} mode={mode}>
-      <BookingsSummaryScreen offer={offer} />
+      <IndividualOfferBooklngsSummaryScreen offer={offer} />
       <ActionBar step={OFFER_WIZARD_STEP_IDS.SUMMARY} isDisabled={false} />
     </IndivualOfferLayout>
   )
@@ -27,4 +28,4 @@ const BookingsSummary = (): JSX.Element | null => {
 
 // Below exports are used by react-router-dom
 // ts-unused-exports:disable-next-line
-export const Component = BookingsSummary
+export const Component = IndividualOfferBookings
